Add a simple authentification scheme
The current scheme is based on basic HTTP Authentification, which is not
very secure as the password is sent on the network in cleartext.

The username / password values provided are then used to connect to the
Database.

This needs to be further improved by using at least a digest + salt.